spring-boot-v2 generator adjusts and cacheable TestContainer
