+++ generators/spring-boot/templates/src/test/kotlin/_package_/config/TestContainersSpringContextCustomizerFactory.kt.ejs
@@ -270,4 +270,9 @@ class TestContainersSpringContextCustomizerFactory: ContextCustomizerFactory {
         }
     }
 
+    override fun hashCode() = TestContainersSpringContextCustomizerFactory::class.java.name.hashCode()
+
+    override fun equals(other: Any?): Boolean {
+        return this.hashCode() == obj.hashCode();
+    }
 }
